refactor: changed overlay modal from route to component
